@@ -3,6 +3,6 @@ package ua.com.aleev.island.action;
 import ua.com.aleev.island.entity.map.Location;
 
 @FunctionalInterface
-public interface Reproducible {
-    void spawn(Location currentLocation);
+public interface Eatable {
+    boolean eat(Location currentLocation);
 }
