@@ -2,6 +2,7 @@ package ua.com.aleev.island.action;
 
 import ua.com.aleev.island.entity.map.Location;
 
+@FunctionalInterface
 public interface Movable {
-   Location move(Location startLocation);
+   void move(Location startLocation);
 }
