@@ -1,12 +1,9 @@
 package ua.com.aleev.island.entity.organism.animal.herbivore;
 
-import ua.com.aleev.island.annotation.OrganismTypeData;
+import ua.com.aleev.island.constant.Sex;
 import ua.com.aleev.island.entity.organism.Limit;
-
-@OrganismTypeData(name = "Caterpillar", icon = "\uD83D\uDC1B", maxWeight = 0.01, maxCount = 1000, maxSpeed = 0, maxFood = 0)
-public class Caterpillar extends Herbivore {
-    public Caterpillar(String name, String icon, double weight, Limit limit) {
-        super(name, icon, weight, limit);
+public class Caterpillar extends Herbivore{
+    public Caterpillar(String name, String icon, double weight, Limit limit, Sex sex) {
+        super(name, icon, weight, limit, sex);
     }
-
 }
